requirement.txt
SQLAlchemy==2.0.25
pandas==2.1.4
scikit-learn==1.3.2
Pillow==10.2.0
streamlit
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-experimental
google-generativeai
faiss-cpu
neo4j
numpy
langgraph
openpyxl==3.1.2
google-search-results
langsmith